Updates to work with ESLint 9
