+++ .github/workflows/test.yml
@@ -15,6 +15,6 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
         with:
-          node-version: '16'
+          node-version: '22'
       - run: npm ci
       - run: npm test
