+++ eslint.config.js
@@ -1,0 +1,3 @@
+import react from './react.js';
+
+export default [...react];
