--- examples/base/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": "../../index.cjs" // this would be "tschaub" in your config
-}
