+++ examples/base/eslint.config.js
@@ -1,0 +1,3 @@
+import base from '../../index.js';
+
+export default [...base];
