--- examples/react/component.js
@@ -1,5 +1,0 @@
-import Message from './Message.js';
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-ReactDOM.render(<Message name="John" />, document.getElementById('root'));
