+++ package.json
@@ -3,13 +3,14 @@
   "version": "14.1.2",
   "description": "A shareable ESLint configuration",
   "type": "module",
+  "main": "index.js",
   "exports": {
-    ".": "./index.cjs",
-    "./react": "./react.cjs"
+    ".": "./index.js",
+    "./react": "./react.js"
   },
   "files": [
-    "index.cjs",
-    "react.cjs"
+    "index.js",
+    "react.js"
   ],
   "repository": {
     "type": "git",
@@ -21,28 +22,28 @@
   "scripts": {
     "lint": "eslint .",
     "test": "npm run check-base && npm run check-react && npm run lint",
-    "check-base": "eslint-config-prettier index.cjs",
-    "check-react": "eslint-config-prettier react.cjs",
+    "check-base": "eslint-config-prettier index.js",
+    "check-react": "eslint-config-prettier react.js",
     "fix-examples": "eslint --fix examples"
   },
   "license": "MIT",
-  "devDependencies": {
-    "eslint": "^8.8.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  },
   "eslintConfig": {
-    "extends": "./index.cjs"
+    "extends": "./index.js"
+  },
+  "devDependencies": {
+    "eslint": "^9.15.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "dependencies": {
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsdoc": ">=38",
-    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsdoc": "^50.5.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
-    "prettier": "^2.7.1"
+    "prettier": "^3.3.3"
   }
 }
